requirements.txt
beautifulsoup4==4.7.1
Django==2.2.4
django-auth-ldap==2.0.0
django-elasticsearch-dsl==0.5.0
django-getenv==1.3.2
django-q==1.0.2
django-widget-tweaks==1.4.5
elasticsearch-dsl==6.1.0
psycopg2==2.8.3
redis==3.3.8
requests==2.22.0
sentry_sdk==0.10.2
